feat(auth): add forgot password link to login screen and implement reset password functionality
feat(categories): update category creation to use iconName instead of imageUrl

fix(inventory): change quantity fields to accept decimal values for better accuracy

feat(stores): replace phone number field with description in store creation and editing

fix(products): update stock fields to handle decimal values for current and minimum stock

feat(users): enhance user role dropdowns with consistent styling

refactor(models): change stock-related fields from int to double for better precision

refactor(auth): remove legacy API service references and streamline token handling

fix(providers): update inventory and store providers to accommodate changes in stock data types

feat(router): add routes for forgot password and reset password screens

+++ lib/features/admin/procurement/widgets/create_procurement_dialog.dart
@@ -519,7 +519,7 @@ class _AddProcurementItemDialogState extends State<AddProcurementItemDialog> {
   final _unitCostController = TextEditingController();
   String? _selectedProductId;
   String? _selectedProductName;
-  int? _availableStock;
+  double? _availableStock;
   String? _stockUnit;
 
   @override
